chore: move functions to utils.rs or target_kind.rs
